[doc] reorganizing the project to make it public
