--- preprocess/requirements.txt
@@ -1,12 +1,0 @@
-#dev
-'numpy >= 1.7.0',
-'nltk >= 3.1',
-'nose >= 1.3.7',
-'pdfminer>=20191020',
-'PyPDF2>=1.26.0'
-
-#doc
-sphinx-rtd-theme
-sphinxcontrib-napoleon
-sphinxcontrib-spelling
-recommonmark
