+++ tests/test_symbols.txt
